backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-dotenv==1.0.1
greenlet==3.1.1
livekit>=1.0
livekit-agents[openai,deepgram,cartesia,silero,turn_detector]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv~=1.0